Refactor into individual modules
