+++ packages/perspective-viewer-datagrid/src/config/cjs.config.js
@@ -3,7 +3,7 @@ const common = require("@finos/perspective/src/config/common.config.js");
 
 module.exports = common({}, config =>
     Object.assign(config, {
-        entry: "./dist/esm/datagrid.js",
+        entry: "./dist/esm/index.js",
         externals: [/^[a-z0-9@]/],
         output: {
             filename: "perspective-viewer-datagrid.js",
