+++ packages/perspective-viewer-datagrid/src/config/umd.config.js
@@ -3,7 +3,7 @@ const common = require("@finos/perspective/src/config/common.config.js");
 
 module.exports = common({}, config =>
     Object.assign(config, {
-        entry: "./dist/esm/datagrid.js",
+        entry: "./dist/esm/index.js",
         output: {
             filename: "perspective-viewer-datagrid.js",
             library: "perspective-viewer-datagrid",
